Accidentally had BIGINT for memory, now INT

+++ database/setup.py
@@ -10,7 +10,7 @@ tables[0] = """CREATE TABLE IF NOT EXISTS machines (
     id INT PRIMARY KEY AUTO_INCREMENT,
     name TEXT NOT NULL,
     cores INT,
-    memory BIGINT
+    memory INT
     );"""
 
 tables[1] = '''CREATE TABLE IF NOT EXISTS benchmarks (
